refactor: extract type-aware exporter args normalization into shared support module
The broker's ExporterConfiguration.normalizeKeys machinery moves verbatim into the new
internal module zeebe-exporter-config-support, and the broker delegates to it. No behavior
change. This gives the normalization semantics a single owner so the upcoming per-physical-tenant
ExporterConfigMerger implementations (ADR-0008 §4) cannot drift from what the broker does when
instantiating an exporter's typed configuration.
