+++ parent/pom.xml
@@ -481,6 +481,12 @@
 
       <dependency>
         <groupId>io.camunda</groupId>
+        <artifactId>zeebe-exporter-config-support</artifactId>
+        <version>${project.version}</version>
+      </dependency>
+
+      <dependency>
+        <groupId>io.camunda</groupId>
         <artifactId>zeebe-exporter-filter</artifactId>
         <version>${project.version}</version>
       </dependency>
