+++ zeebe/broker/src/main/java/io/camunda/zeebe/broker/exporter/context/ExporterConfiguration.java
@@ -14,7 +14,6 @@ import com.fasterxml.jackson.core.json.JsonReadFeature;
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.DeserializationContext;
 import com.fasterxml.jackson.databind.DeserializationFeature;
-import com.fasterxml.jackson.databind.JavaType;
 import com.fasterxml.jackson.databind.JsonDeserializer;
 import com.fasterxml.jackson.databind.JsonSerializer;
 import com.fasterxml.jackson.databind.MapperFeature;
@@ -27,15 +26,12 @@ import com.fasterxml.jackson.databind.util.TokenBuffer;
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
 import io.camunda.zeebe.exporter.api.context.Configuration;
 import io.camunda.zeebe.exporter.api.context.StrictConfiguration;
+import io.camunda.zeebe.exporter.support.ExporterConfigMergeSupport;
 import io.camunda.zeebe.util.ReflectUtil;
 import java.io.IOException;
-import java.lang.reflect.Array;
 import java.nio.file.Path;
 import java.time.Duration;
-import java.util.ArrayList;
-import java.util.LinkedHashMap;
 import java.util.List;
-import java.util.Locale;
 import java.util.Map;
 import java.util.function.Consumer;
 import org.slf4j.Logger;
@@ -109,8 +105,10 @@ public record ExporterConfiguration(String id, Map<String, Object> arguments)
       // Normalize all keys to lowercase-no-separator so that camelCase, kebab-case, and the
       // all-lowercase concatenated form produced by Spring Boot env var lowercasing are all
       // equivalent. ACCEPT_CASE_INSENSITIVE_PROPERTIES then matches the normalised key to the
-      // Java field name regardless of original casing.
-      final var normalized = normalizeKeys(values, MAPPER.constructType(configClass));
+      // Java field name regardless of original casing. The type-aware walk is shared with the
+      // per-physical-tenant ExporterConfigMerger implementations (single owner of the
+      // normalization semantics, ADR-0008 §4).
+      final var normalized = ExporterConfigMergeSupport.normalize(configClass, values);
       final var mapper =
           configClass.isAnnotationPresent(StrictConfiguration.class) ? STRICT_MAPPER : MAPPER;
       try {
@@ -143,148 +141,6 @@ public record ExporterConfiguration(String id, Map<String, Object> arguments)
     } else {
       return ReflectUtil.newInstance(configClass);
     }
-  }
-
-  /**
-   * Normalises keys that represent configuration properties to lowercase with separators stripped,
-   * so that {@code myField}, {@code my-field}, {@code MY-FIELD}, and {@code myfield} all map to the
-   * same canonical key.
-   *
-   * <p>The normalisation is type-aware: object-property maps are normalised recursively, while map
-   * keys of {@code Map<K, V>} fields are preserved as user data.
-   *
-   * @param map the raw exporter args map
-   * @return the same structure with normalised lowercase keys
-   */
-  private static Map<String, Object> normalizeKeys(
-      final Map<String, Object> map, final JavaType targetType) {
-    final var result = new LinkedHashMap<String, Object>(map.size());
-    final var propertyTypesByKey = propertyTypesByNormalizedKey(targetType);
-    final var indexedElementType = indexedElementType(targetType, map);
-    for (final var entry : map.entrySet()) {
-      final var normalizedKey = normalizeKey(entry.getKey());
-      var propertyType = propertyTypesByKey.get(normalizedKey);
-      if (propertyType == null) {
-        propertyType = indexedElementType;
-      }
-      result.put(normalizedKey, normalizeValue(entry.getValue(), propertyType));
-    }
-    return result;
-  }
-
-  private static JavaType indexedElementType(
-      final JavaType targetType, final Map<String, Object> map) {
-    if (targetType == null || (!targetType.isCollectionLikeType() && !targetType.isArrayType())) {
-      return null;
-    }
-
-    var hasNumericKeys = false;
-    var hasNonNumericKeys = false;
-    for (final var key : map.keySet()) {
-      try {
-        Integer.parseInt(key);
-        hasNumericKeys = true;
-      } catch (final NumberFormatException e) {
-        hasNonNumericKeys = true;
-      }
-    }
-
-    if (hasNumericKeys && hasNonNumericKeys) {
-      throw new IllegalArgumentException(
-          "Cannot mix indexed (numeric) and named keys in configuration map targeting a collection: "
-              + map.keySet());
-    }
-
-    return hasNumericKeys ? targetType.getContentType() : null;
-  }
-
-  private static Map<String, JavaType> propertyTypesByNormalizedKey(final JavaType targetType) {
-    if (targetType == null || targetType.isMapLikeType() || targetType.isContainerType()) {
-      return Map.of();
-    }
-
-    final var properties =
-        MAPPER.getDeserializationConfig().introspect(targetType).findProperties();
-    if (properties.isEmpty()) {
-      return Map.of();
-    }
-
-    final var propertyTypes = new LinkedHashMap<String, JavaType>(properties.size());
-    for (final var property : properties) {
-      final var propertyType = property.getPrimaryType();
-      if (propertyType != null) {
-        propertyTypes.put(normalizeKey(property.getName()), propertyType);
-      }
-    }
-    return propertyTypes;
-  }
-
-  private static Object normalizeValue(final Object value, final JavaType targetType) {
-    if (value instanceof final Map<?, ?> nestedMap) {
-      return normalizeMapValue(nestedMap, targetType);
-    }
-
-    if (value instanceof final Iterable<?> iterable) {
-      return normalizeIterableValue(iterable, targetType);
-    }
-
-    if (value != null && value.getClass().isArray()) {
-      return normalizeArrayValue(value, targetType);
-    }
-
-    return value;
-  }
-
-  private static Object normalizeMapValue(final Map<?, ?> nestedMap, final JavaType targetType) {
-    if (targetType == null || targetType.hasRawClass(Object.class)) {
-      return nestedMap;
-    }
-
-    if (targetType.isMapLikeType()) {
-      return normalizeMapValues(nestedMap, targetType.getContentType());
-    }
-
-    @SuppressWarnings("unchecked")
-    final var typedNested = (Map<String, Object>) nestedMap;
-    return normalizeKeys(typedNested, targetType);
-  }
-
-  private static List<Object> normalizeIterableValue(
-      final Iterable<?> iterable, final JavaType targetType) {
-    final var contentType =
-        targetType != null && targetType.isCollectionLikeType()
-            ? targetType.getContentType()
-            : null;
-    final var normalized = new ArrayList<>();
-    for (final var item : iterable) {
-      normalized.add(normalizeValue(item, contentType));
-    }
-    return normalized;
-  }
-
-  private static Object[] normalizeArrayValue(final Object value, final JavaType targetType) {
-    final var length = Array.getLength(value);
-    final var contentType =
-        targetType != null && targetType.isArrayType() ? targetType.getContentType() : null;
-    final var normalized = new Object[length];
-    for (int i = 0; i < length; i++) {
-      normalized[i] = normalizeValue(Array.get(value, i), contentType);
-    }
-    return normalized;
-  }
-
-  private static Map<Object, Object> normalizeMapValues(
-      final Map<?, ?> map, final JavaType contentType) {
-    final var result = new LinkedHashMap<Object, Object>(map.size());
-    for (final var entry : map.entrySet()) {
-      result.put(entry.getKey(), normalizeValue(entry.getValue(), contentType));
-    }
-    return result;
-  }
-
-  /** Strips hyphens and lowercases a key so all naming styles map to the same canonical form. */
-  private static String normalizeKey(final String name) {
-    return name.replace("-", "").toLowerCase(Locale.ROOT);
   }
 
   /**
